# docs/swagger/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Vektonn API V1",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "v1"
  },
  "paths": {
    "/api/v1/search/{indexName}/{indexVersion}": {
      "get": {
        "tags": [
          "SearchApi"
        ],
        "parameters": [
          {
            "name": "indexName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "indexVersion",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchQueryDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchQueryDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SearchQueryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchResultDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchResultDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchResultDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "SearchApi"
        ],
        "parameters": [
          {
            "name": "indexName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "indexVersion",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchQueryDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchQueryDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SearchQueryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchResultDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchResultDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchResultDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/upload/{dataSourceName}/{dataSourceVersion}": {
      "post": {
        "tags": [
          "UploadApi"
        ],
        "parameters": [
          {
            "name": "dataSourceName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dataSourceVersion",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/InputDataPointDto"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/InputDataPointDto"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/InputDataPointDto"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AttributeDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "value": {
            "$ref": "#/components/schemas/AttributeValueDto"
          }
        },
        "additionalProperties": false
      },
      "AttributeValueDto": {
        "type": "object",
        "properties": {
          "string": {
            "type": "string",
            "nullable": true
          },
          "guid": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "bool": {
            "type": "boolean",
            "nullable": true
          },
          "int64": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "dateTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DenseVectorDto": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/VectorDto"
          }
        ],
        "additionalProperties": false
      },
      "FoundDataPointDto": {
        "type": "object",
        "properties": {
          "vector": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/DenseVectorDto"
              },
              {
                "$ref": "#/components/schemas/SparseVectorDto"
              }
            ]
          },
          "attributes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AttributeDto"
            }
          },
          "distance": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "InputDataPointDto": {
        "type": "object",
        "properties": {
          "attributes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AttributeDto"
            }
          },
          "vector": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/DenseVectorDto"
              },
              {
                "$ref": "#/components/schemas/SparseVectorDto"
              }
            ],
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "SearchQueryDto": {
        "type": "object",
        "properties": {
          "splitFilter": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AttributeDto"
            },
            "nullable": true
          },
          "queryVectors": {
            "type": "array",
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/DenseVectorDto"
                },
                {
                  "$ref": "#/components/schemas/SparseVectorDto"
                }
              ]
            }
          },
          "k": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "SearchResultDto": {
        "type": "object",
        "properties": {
          "queryVector": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/DenseVectorDto"
              },
              {
                "$ref": "#/components/schemas/SparseVectorDto"
              }
            ]
          },
          "nearestDataPoints": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FoundDataPointDto"
            }
          }
        },
        "additionalProperties": false
      },
      "SparseVectorDto": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/VectorDto"
          }
        ],
        "properties": {
          "coordinateIndices": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          }
        },
        "additionalProperties": false
      },
      "VectorDto": {
        "type": "object",
        "properties": {
          "isSparse": {
            "type": "boolean"
          },
          "coordinates": {
            "type": "array",
            "items": {
              "type": "number",
              "format": "double"
            }
          }
        },
        "additionalProperties": false
      }
    }
  }
}